requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.4
plotly==5.15.0
gunicorn==20.1.0